pyproject.toml
[project]
name = "transportation_flow"
version = "0.1.0"
description = "transportation_flow using crewAI"
authors = [
    { name = "Your Name", email = "you@example.com" },
]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.126.0,<1.0.0",
    "fastapi>=0.115.12",
    "langchain>=0.3.25",
    "openai>=1.75.0",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "redis>=6.2.0",
    "requests>=2.32.3",
    "uvicorn>=0.34.3",
]

[project.scripts]
kickoff = "transportation_flow.main:kickoff"
run_crew = "transportation_flow.main:kickoff"
plot = "transportation_flow.main:plot"

[build-system]
requires = [
    "hatchling",
]
build-backend = "hatchling.build"
